// vestara-apps/api-builder/src/api/contracts.ts
export interface ApiField {
  readonly id: string;
  readonly name: string;
  readonly type: string;
  readonly required?: boolean;
  readonly unique?: boolean;
  readonly indexed?: boolean;
  readonly enumValues?: readonly string[];
}